edit attr tabpane
